Added option to rebuild from scratch in setup_env

+++ setup_env.bash
@@ -1,5 +1,35 @@
 #!/bin/bash
-rm -rf .venv
+
+HELP_MESSAGE="Usage: ./build.sh [-r, --rebuild]
+Build the '${IMAGE_NAME}' image.
+Options:
+  -r, --rebuild            Remove the previous venv completely before rebuilding
+  -h, --help               Show this help message."
+
+# Parse input flags
+BUILD_FLAGS=()
+while [ "$#" -gt 0 ]; do
+  case "$1" in
+  -r | --rebuild)
+    rebuild=true
+    shift 1
+    ;;
+  -h | --help)
+    echo "${HELP_MESSAGE}"
+    exit 0
+    ;;
+  *)
+    echo "Unknown option: $1" >&2
+    exit 1
+    ;;
+  esac
+done
+
+# Check if we need to remove past files and caches
+if [ "$rebuild"=true ] ; then
+    rm -rf .venv
+fi
+
 python3.9 -m venv .venv          # Create venv (use any python version you like >=3.9)
 source .venv/bin/activate        # Activate venv
 pip install -r requirements.txt  # Install package requirements
